requirements.txt
Flask
twilio
python-dotenv
pymongo
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
dnspython
cryptography>=42.0.0